pyproject.toml
[tool.poetry]
name = "glovo-api-python"
version = "1.0.0"
description = ""
authors = ["SoftButterfly Development Team <dev@softbutterfly.io>"]
license = "BSD 3-Clause License"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.22.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.4.4"
black = "^19.10b0"
codecov = "^2.0.15"
coverage = "^5.0.3"
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
jupyterlab = "^1.2.5"
mypy = "^0.761"
pre-commit = "^1.21.0"
pydocstyle = "^5.0.2"
pylint = "^2.4.4"
pytest = "^5.3.3"
pytest-cov = "^2.8.1"
pytest-vcr = "^1.0.2"
python-dotenv = "^0.10.4"
tox = "^3.14.3"
twine = "^3.1.1"

[tool.black]
target_version = ['py35', 'py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |_build
  |buck-out
  |build
  |dist
  |legacy
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
